Update BIP32.swift to make `privKey` public.
In order to use individual child keys to sign a PSBT or raw transaction we need to be able to access the `privKey`. Considering the `xprv` property is public it does not seem to do any harm to expose a child key. Signing psbt's with anything but the root xprv does not seem to work. I would like to solve that issue but in order to do that I need to be able to access individual child private keys. Generally speaking it would be nice to access indvidual private keys to sign transactions with irregardless.